refactor(analytics): rename posthog.api → analytics.service to drop provider name from module
