+++ packages/dashboard/src/features/analytics/hooks/useWebOverview.ts
@@ -1,11 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
 import type { PosthogTimeframe } from '@insforge/shared-schemas';
-import { posthogApi } from '#features/analytics/services/posthog.api';
+import { analyticsService } from '#features/analytics/services/analytics.service';
 
 export function useWebOverview(timeframe: PosthogTimeframe, enabled: boolean) {
   return useQuery({
     queryKey: ['posthog', 'web-overview', timeframe],
-    queryFn: () => posthogApi.getWebOverview(timeframe),
+    queryFn: () => analyticsService.getWebOverview(timeframe),
     enabled,
     staleTime: 60_000,
   });
