+++ packages/dashboard/src/features/analytics/hooks/useWebStats.ts
@@ -1,11 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
 import type { PosthogTimeframe } from '@insforge/shared-schemas';
-import { posthogApi, type Breakdown } from '#features/analytics/services/posthog.api';
+import { analyticsService, type Breakdown } from '#features/analytics/services/analytics.service';
 
 export function useWebStats(breakdown: Breakdown, timeframe: PosthogTimeframe, enabled: boolean) {
   return useQuery({
     queryKey: ['posthog', 'web-stats', breakdown, timeframe],
-    queryFn: () => posthogApi.getWebStats(breakdown, timeframe),
+    queryFn: () => analyticsService.getWebStats(breakdown, timeframe),
     enabled,
     staleTime: 60_000,
   });
